Move detect code out of frame.js for now
