+++ data/frame.js
@@ -1,8 +1,5 @@
 /* global addMessageListener, removeMessageListener, sendAsyncMessage, content */
 
-Components.utils.import("resource://gre/modules/Services.jsm");
-Components.utils.import("resource://gre/modules/XPCOMUtils.jsm");
-
 let muting = false;
 let tabMuted = false;
 let previous = false;
@@ -17,46 +14,6 @@ addMessageListener("NoiseControl:checkPlugins", checkPlugins);
 addMessageListener("NoiseControl:mute", muteListener);
 addMessageListener("NoiseControl:disable", disableListener);
 checkNoise();
-
-let s = new Set();
-let observer = {
-	observe: function(subject, topic, data) {
-		if (!subject) {
-			// console.log("eh?!");
-			return;
-		}
-		// console.log("this script: " + (subject.top == content) + ", is top:" + (subject == content));
-		if (subject.top == content) {
-			// console.log(topic, data);
-			if (data == "active") {
-				s.add(subject);
-				subject.addEventListener("unload", this.windowUnload);
-			} else {
-				s.delete(subject);
-			}
-			try {
-				sendAsyncMessage("NoiseControl:testMessage", s.size);
-			} catch(ex) {
-				// console.error(ex);
-			}
-		}
-	},
-	windowUnload: function(event) {
-		s.delete(event.currentTarget);
-		try {
-			sendAsyncMessage("NoiseControl:testMessage", s.size);
-		} catch(ex) {
-			// console.error(ex);
-		}
-	},
-	QueryInterface: XPCOMUtils.generateQI([
-		Components.interfaces.nsIObserver,
-		Components.interfaces.nsISupportsWeakReference,
-		Components.interfaces.nsISupports
-	])
-};
-Services.obs.addObserver(observer, "media-playback", true);
-// console.log("observer added");
 
 function checkNoise() {
 	if (muting) {
@@ -201,6 +158,4 @@ function disableListener()  {
 	removeMessageListener("NoiseControl:checkNoise", forceCheckNoise);
 	removeMessageListener("NoiseControl:mute", muteListener);
 	removeMessageListener("NoiseControl:disable", disableListener);
-
-	Services.obs.removeObserver(observer, "media-playback");
 }
