+++ lib/main.js
@@ -7,14 +7,15 @@ const { Cc, Ci, Cu } = require("chrome");
 const { data } = require("sdk/self");
 const { viewFor } = require("sdk/view/core");
 const preferences = require("sdk/simple-prefs");
+const prefService = require("sdk/preferences/service");
 
 const FRAME_SCRIPT_URI = data.url("frame.js");
 const CSS_PI_DATA = "href=\"" + data.url("noise-control.css") + "\" type=\"text/css\"";
 
 Cu.import("resource://gre/modules/Services.jsm");
 
-preferences.set("media.defaultAudioChannel", "content");
-preferences.set("media.useAudioChannelService", true);
+prefService.set("media.defaultAudioChannel", "content");
+prefService.set("media.useAudioChannelService", true);
 
 let messageManager = Cc["@mozilla.org/globalmessagemanager;1"].getService(Ci.nsIMessageListenerManager);
 let listener = {
@@ -23,8 +24,7 @@ let listener = {
 		"NoiseControl:hasPlugins",
 		"NoiseControl:unloaded",
 		"NoiseControl:unmuted",
-		"PluginContent:UpdateHiddenPluginUI",
-		// "NoiseControl:testMessage"
+		"PluginContent:UpdateHiddenPluginUI"
 	],
 	init: function() {
 		for (let m of this.messages) {
@@ -68,10 +68,6 @@ let listener = {
 		case "PluginContent:UpdateHiddenPluginUI":
 			browser.messageManager.sendAsyncMessage("NoiseControl:checkPlugins");
 			break;
-		// case "NoiseControl:testMessage":
-		// 	let hasNoise = Number(message.data);
-		// 	indicator.classList[hasNoise ? "add" : "remove"]("noisy");
-		// 	break;
 		default:
 			// console.log(message.name, JSON.stringify(message.json));
 			break;
@@ -101,8 +97,8 @@ exports.onUnload = function(reason) {
 	listener.cleanup();
 	preferences.removeListener("indicator.colour", indicatorColourChanged);
 
-	preferences.reset("media.defaultAudioChannel", "content");
-	preferences.reset("media.useAudioChannelService");
+	prefService.reset("media.defaultAudioChannel", "content");
+	prefService.reset("media.useAudioChannelService");
 
 	for (let sdkWindow of require("sdk/windows").browserWindows) {
 		let chromeWindow = viewFor(sdkWindow);
